finalised driver dashboard
added a sync orders button to pickupRequest page where it checks for orders placed on the current day and loads corresponding pickup requests into the pickup request section for the driver to accept or decline.

removed earnings report tab as it's not needed.

updated logout button colour to match on all tabs of driver dashboard.

updated server.py accordingly to make everything function correctly.

also added a feature for when the driver accepts a pickup, a pickup time of 30mins, from then moment of acceptance, is automatically added.